REF, CLN: Moved common ingest utils to utils.py and refactored old utils.py to s3_sync and s3_async.py
